requirements.txt
telepot==8.1
pyserial==3.1
